add nft test to cast, update and relinquish
