+++ cNft-Governance/governance-program-library/programs/nft-voter/tests/create_max_voter_weight_record.rs
@@ -5,7 +5,6 @@ use solana_program_test::*;
 use solana_sdk::transport::TransportError;
 
 mod program_test;
-mod nft_test;
 
 #[tokio::test]
 async fn test_create_max_voter_weight_record() -> Result<(), TransportError> {
