+++ cNft-Governance/governance-program-library/programs/nft-voter/tests/create_nft_weight_record.rs
@@ -270,3 +270,55 @@ async fn test_create_nft_weight_record_with_invalid_collection_error() -> Result
     assert_nft_voter_err(err, NftVoterError::CollectionNotFound);
     Ok(())
 }
+
+#[tokio::test]
+async fn test_create_nft_weight_record_with_no_nft_error() -> Result<(), TransportError> {
+    // Arrange
+    let mut nft_voter_test = NftVoterTest::start_new().await;
+    let realm_cookie = nft_voter_test.governance.with_realm().await?;
+    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
+    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
+    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
+        &registrar_cookie
+    ).await?;
+
+    nft_voter_test.with_collection(
+        &registrar_cookie,
+        &nft_collection_cookie,
+        &max_voter_weight_record_cookie,
+        Some(ConfigureCollectionArgs {
+            weight: 10,
+            size: 20,
+        })
+    ).await?;
+
+    let voter_cookie = nft_voter_test.bench.with_wallet().await;
+
+    let voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
+        &registrar_cookie,
+        &voter_cookie
+    ).await?;
+
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        Some(CreateNftArgs {
+            amount: 0,
+            ..Default::default()
+        })
+    ).await?;
+
+    // Act
+    let err = nft_voter_test
+        .with_create_nft_weight_record(
+            &registrar_cookie,
+            &voter_weight_record_cookie,
+            &voter_cookie,
+            &[&nft_cookie]
+        ).await
+        .err()
+        .unwrap();
+
+    assert_nft_voter_err(err, NftVoterError::InvalidNftAmount);
+    Ok(())
+}
