--- cNft-Governance/governance-program-library/programs/nft-voter/tests/nft_test/mod.rs
@@ -1,3 +1,0 @@
-mod cast_nft_vote;
-mod relinquish_nft_vote;
-mod update_nft_voter_weight_record;
