--- cNft-Governance/governance-program-library/programs/nft-voter/tests/nft_test/test_nft_mint.rs
@@ -1,6 +1,0 @@
-use solana_program_test::*;
-
-#[tokio::test]
-async fn test_nft_mint() -> Result<(), TransportError> {
-    Ok(())
-}
