--- cNft-Governance/governance-program-library/programs/nft-voter/tests/nft_test/update_nft_voter_weight_record.rs
@@ -1,548 +1,0 @@
-use crate::program_test::nft_voter_test::ConfigureCollectionArgs;
-use gpl_nft_voter::error::NftVoterError;
-use gpl_nft_voter::state::*;
-use crate::program_test::nft_voter_test::NftVoterTest;
-use crate::program_test::token_metadata_test::CreateNftArgs;
-use crate::program_test::tools::*;
-use solana_program_test::*;
-use solana_sdk::transport::TransportError;
-
-// mod program_test;
-
-#[tokio::test]
-async fn test_update_nft_voter_weight_record() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    let _collection_config_cookie = nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    nft_voter_test.bench.advance_clock().await;
-    let clock = nft_voter_test.bench.get_clock().await;
-
-    // Act
-    nft_voter_test.update_nft_voter_weight_record(
-        &registrar_cookie,
-        &mut voter_weight_record_cookie,
-        VoterWeightAction::CreateProposal,
-        &[&nft1_cookie]
-    ).await?;
-
-    // Assert
-
-    let voter_weight_record = nft_voter_test.get_voter_weight_record(
-        &voter_weight_record_cookie.address
-    ).await;
-
-    assert_eq!(voter_weight_record.voter_weight, 10);
-    assert_eq!(voter_weight_record.voter_weight_expiry, Some(clock.slot));
-    assert_eq!(voter_weight_record.weight_action, Some(VoterWeightAction::CreateProposal.into()));
-    assert_eq!(voter_weight_record.weight_action_target, None);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_multiple_nfts() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    let _collection_config_cookie = nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft_cookie1 = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    let nft_cookie2 = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    nft_voter_test.bench.advance_clock().await;
-    let clock = nft_voter_test.bench.get_clock().await;
-
-    // Act
-    nft_voter_test.update_nft_voter_weight_record(
-        &registrar_cookie,
-        &mut voter_weight_record_cookie,
-        VoterWeightAction::CreateProposal,
-        &[&nft_cookie1, &nft_cookie2]
-    ).await?;
-
-    // Assert
-
-    let voter_weight_record = nft_voter_test.get_voter_weight_record(
-        &voter_weight_record_cookie.address
-    ).await;
-
-    assert_eq!(voter_weight_record.voter_weight, 20);
-    assert_eq!(voter_weight_record.voter_weight_expiry, Some(clock.slot));
-    assert_eq!(voter_weight_record.weight_action, Some(VoterWeightAction::CreateProposal.into()));
-    assert_eq!(voter_weight_record.weight_action_target, None);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CastVote,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::CastVoteIsNotAllowed);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_unverified_collection_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    // Create NFT without verified collection
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        Some(CreateNftArgs {
-            verify_collection: false,
-            ..Default::default()
-        })
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateGovernance,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::CollectionMustBeVerified);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_invalid_owner_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let voter_cookie2 = nft_voter_test.bench.with_wallet().await;
-
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie2,
-        None
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateGovernance,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::VoterDoesNotOwnNft);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_invalid_collection_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft_collection_cookie2 = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie2,
-        &voter_cookie,
-        None
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateGovernance,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::CollectionNotFound);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_invalid_metadata_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let mut nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        Some(CreateNftArgs {
-            verify_collection: false,
-            ..Default::default()
-        })
-    ).await?;
-
-    let nft2_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    // Try to use verified NFT Metadata
-    nft1_cookie.metadata = nft2_cookie.metadata;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateGovernance,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::TokenMetadataDoesNotMatch);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_voter_weight_with_same_nft_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        None
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        None
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateProposal,
-            &[&nft_cookie, &nft_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-
-    assert_nft_voter_err(err, NftVoterError::DuplicatedNftDetected);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_update_nft_voter_weight_record_with_no_nft_error() -> Result<(), TransportError> {
-    // Arrange
-    let mut nft_voter_test = NftVoterTest::start_new().await;
-
-    let realm_cookie = nft_voter_test.governance.with_realm().await?;
-
-    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
-
-    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
-
-    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
-        &registrar_cookie
-    ).await?;
-
-    let _collection_config_cookie = nft_voter_test.with_collection(
-        &registrar_cookie,
-        &nft_collection_cookie,
-        &max_voter_weight_record_cookie,
-        Some(ConfigureCollectionArgs {
-            weight: 10,
-            size: 20,
-        })
-    ).await?;
-
-    let voter_cookie = nft_voter_test.bench.with_wallet().await;
-
-    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
-        &registrar_cookie,
-        &voter_cookie
-    ).await?;
-
-    let nft1_cookie = nft_voter_test.token_metadata.with_nft_v2(
-        &nft_collection_cookie,
-        &voter_cookie,
-        Some(CreateNftArgs {
-            amount: 0,
-            ..Default::default()
-        })
-    ).await?;
-
-    // Act
-    let err = nft_voter_test
-        .update_nft_voter_weight_record(
-            &registrar_cookie,
-            &mut voter_weight_record_cookie,
-            VoterWeightAction::CreateProposal,
-            &[&nft1_cookie]
-        ).await
-        .err()
-        .unwrap();
-
-    // Assert
-    assert_nft_voter_err(err, NftVoterError::InvalidNftAmount);
-
-    Ok(())
-}
