+++ cNft-Governance/governance-program-library/programs/nft-voter/tests/update_voter_weight_record.rs
@@ -6,7 +6,68 @@ use crate::program_test::tools::assert_nft_voter_err;
 mod program_test;
 
 #[tokio::test]
-async fn test_update_voter_weight_record() -> Result<(), TransportError> {
+async fn test_update_voter_weight_record_with_nft() -> Result<(), TransportError> {
+    let mut nft_voter_test = NftVoterTest::start_new().await;
+    let realm_cookie = nft_voter_test.governance.with_realm().await?;
+    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
+    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
+        &registrar_cookie
+    ).await?;
+    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
+
+    nft_voter_test.with_collection(
+        &registrar_cookie,
+        &nft_collection_cookie,
+        &max_voter_weight_record_cookie,
+        Some(ConfigureCollectionArgs {
+            weight: 3,
+            size: 11,
+        })
+    ).await?;
+
+    let voter_cookie = nft_voter_test.bench.with_wallet().await;
+    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
+        &registrar_cookie,
+        &voter_cookie
+    ).await?;
+
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
+    nft_voter_test.bench.advance_clock().await;
+    let clock = nft_voter_test.bench.get_clock().await;
+
+    let nft_weight_record_cookies = nft_voter_test.with_create_nft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&nft_cookie]
+    ).await?;
+
+    nft_voter_test.update_voter_weight_record(
+        &registrar_cookie,
+        &mut voter_weight_record_cookie,
+        VoterWeightAction::CreateProposal,
+        &nft_weight_record_cookies.iter().collect::<Vec<_>>()
+    ).await?;
+
+    let voter_weight_record = nft_voter_test.get_voter_weight_record(
+        &voter_weight_record_cookie.address
+    ).await;
+
+    assert_eq!(voter_weight_record.voter_weight, 3);
+    assert_eq!(voter_weight_record.voter_weight_expiry, Some(clock.slot));
+    assert_eq!(voter_weight_record.weight_action, Some(VoterWeightAction::CreateProposal.into()));
+    assert_eq!(voter_weight_record.weight_action_target, None);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_update_voter_weight_record_with_cnft() -> Result<(), TransportError> {
     let mut nft_voter_test = NftVoterTest::start_new().await;
     let realm_cookie = nft_voter_test.governance.with_realm().await?;
     let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
@@ -80,7 +141,160 @@ async fn test_update_voter_weight_record() -> Result<(), TransportError> {
 }
 
 #[tokio::test]
+async fn test_update_voter_weight_record_with_nft_and_cnft() -> Result<(), TransportError> {
+    let mut nft_voter_test = NftVoterTest::start_new().await;
+    let realm_cookie = nft_voter_test.governance.with_realm().await?;
+    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
+    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
+        &registrar_cookie
+    ).await?;
+    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
+
+    nft_voter_test.with_collection(
+        &registrar_cookie,
+        &nft_collection_cookie,
+        &max_voter_weight_record_cookie,
+        Some(ConfigureCollectionArgs {
+            weight: 3,
+            size: 11,
+        })
+    ).await?;
+
+    let voter_cookie = nft_voter_test.bench.with_wallet().await;
+    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
+        &registrar_cookie,
+        &voter_cookie
+    ).await?;
+
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
+    // mint compressed nft
+    let mut tree_cookie = nft_voter_test.merkle_tree.with_merkle_tree(None).await?;
+    let leaf_cookie = nft_voter_test.token_metadata.with_compressed_nft_to_collection(
+        &nft_collection_cookie,
+        &mut tree_cookie,
+        &voter_cookie
+    ).await?;
+
+    nft_voter_test.bench.advance_clock().await;
+    let clock = nft_voter_test.bench.get_clock().await;
+
+    let (leaf_verification_cookie, proofs, _) =
+        nft_voter_test.merkle_tree.get_leaf_verification_info(
+            &mut tree_cookie,
+            &leaf_cookie,
+            5,
+            8
+        ).await?;
+
+    let nft_weight_record_cookies = nft_voter_test.with_create_nft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&nft_cookie]
+    ).await?;
+
+    let cnft_weight_record_cookies = nft_voter_test.with_create_cnft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&leaf_cookie],
+        &[&leaf_verification_cookie],
+        &[&proofs]
+    ).await?;
+
+    nft_voter_test.update_voter_weight_record(
+        &registrar_cookie,
+        &mut voter_weight_record_cookie,
+        VoterWeightAction::CreateProposal,
+        &[&nft_weight_record_cookies[0], &cnft_weight_record_cookies[0]]
+    ).await?;
+
+    let voter_weight_record = nft_voter_test.get_voter_weight_record(
+        &voter_weight_record_cookie.address
+    ).await;
+
+    assert_eq!(voter_weight_record.voter_weight, 6);
+    assert_eq!(voter_weight_record.voter_weight_expiry, Some(clock.slot));
+    assert_eq!(voter_weight_record.weight_action, Some(VoterWeightAction::CreateProposal.into()));
+    assert_eq!(voter_weight_record.weight_action_target, None);
+
+    Ok(())
+}
+
+#[tokio::test]
 async fn test_update_voter_weight_record_with_multiple_nfts() -> Result<(), TransportError> {
+    let mut nft_voter_test = NftVoterTest::start_new().await;
+    let realm_cookie = nft_voter_test.governance.with_realm().await?;
+    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
+    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
+        &registrar_cookie
+    ).await?;
+    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
+
+    nft_voter_test.with_collection(
+        &registrar_cookie,
+        &nft_collection_cookie,
+        &max_voter_weight_record_cookie,
+        Some(ConfigureCollectionArgs {
+            weight: 3,
+            size: 11,
+        })
+    ).await?;
+
+    let voter_cookie = nft_voter_test.bench.with_wallet().await;
+    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
+        &registrar_cookie,
+        &voter_cookie
+    ).await?;
+
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
+    let nft_cookie2 = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
+    nft_voter_test.bench.advance_clock().await;
+    let clock = nft_voter_test.bench.get_clock().await;
+
+    let nft_weight_record_cookies = nft_voter_test.with_create_nft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&nft_cookie, &nft_cookie2]
+    ).await?;
+
+    nft_voter_test.update_voter_weight_record(
+        &registrar_cookie,
+        &mut voter_weight_record_cookie,
+        VoterWeightAction::CreateProposal,
+        &nft_weight_record_cookies.iter().collect::<Vec<_>>()
+    ).await?;
+
+    let voter_weight_record = nft_voter_test.get_voter_weight_record(
+        &voter_weight_record_cookie.address
+    ).await;
+
+    assert_eq!(voter_weight_record.voter_weight, 6);
+    assert_eq!(voter_weight_record.voter_weight_expiry, Some(clock.slot));
+    assert_eq!(voter_weight_record.weight_action, Some(VoterWeightAction::CreateProposal.into()));
+    assert_eq!(voter_weight_record.weight_action_target, None);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_update_voter_weight_record_with_multiple_cnfts() -> Result<(), TransportError> {
     let mut nft_voter_test = NftVoterTest::start_new().await;
     let realm_cookie = nft_voter_test.governance.with_realm().await?;
     let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
@@ -193,9 +407,14 @@ async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(
         &voter_cookie
     ).await?;
 
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
     // mint compressed nft
     let mut tree_cookie = nft_voter_test.merkle_tree.with_merkle_tree(None).await?;
-
     let leaf_cookie = nft_voter_test.token_metadata.with_compressed_nft_to_collection(
         &nft_collection_cookie,
         &mut tree_cookie,
@@ -203,6 +422,13 @@ async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(
     ).await?;
 
     nft_voter_test.bench.advance_clock().await;
+
+    let nft_weight_record_cookies = nft_voter_test.with_create_nft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&nft_cookie]
+    ).await?;
 
     let (leaf_verification_cookie, proofs, _) =
         nft_voter_test.merkle_tree.get_leaf_verification_info(
@@ -212,7 +438,7 @@ async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(
             8
         ).await?;
 
-    let nft_weight_record_cookies = nft_voter_test.with_create_cnft_weight_record(
+    let cnft_weight_record_cookies = nft_voter_test.with_create_cnft_weight_record(
         &registrar_cookie,
         &voter_weight_record_cookie,
         &voter_cookie,
@@ -226,7 +452,7 @@ async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(
             &registrar_cookie,
             &mut voter_weight_record_cookie,
             VoterWeightAction::CastVote,
-            &nft_weight_record_cookies.iter().collect::<Vec<_>>()
+            &[&nft_weight_record_cookies[0], &cnft_weight_record_cookies[0]]
         ).await
         .err()
         .unwrap();
@@ -237,6 +463,62 @@ async fn test_update_voter_weight_with_cast_vote_not_allowed_error() -> Result<(
 
 #[tokio::test]
 async fn test_update_voter_weight_with_same_nft_error() -> Result<(), TransportError> {
+    let mut nft_voter_test = NftVoterTest::start_new().await;
+    let realm_cookie = nft_voter_test.governance.with_realm().await?;
+    let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
+    let max_voter_weight_record_cookie = nft_voter_test.with_max_voter_weight_record(
+        &registrar_cookie
+    ).await?;
+    let nft_collection_cookie = nft_voter_test.token_metadata.with_nft_collection(Some(10)).await?;
+
+    nft_voter_test.with_collection(
+        &registrar_cookie,
+        &nft_collection_cookie,
+        &max_voter_weight_record_cookie,
+        Some(ConfigureCollectionArgs {
+            weight: 3,
+            size: 11,
+        })
+    ).await?;
+
+    let voter_cookie = nft_voter_test.bench.with_wallet().await;
+
+    let mut voter_weight_record_cookie = nft_voter_test.with_voter_weight_record(
+        &registrar_cookie,
+        &voter_cookie
+    ).await?;
+
+    let nft_cookie = nft_voter_test.token_metadata.with_nft_v2(
+        &nft_collection_cookie,
+        &voter_cookie,
+        None
+    ).await?;
+
+    nft_voter_test.bench.advance_clock().await;
+
+    let nft_weight_record_cookies = nft_voter_test.with_create_nft_weight_record(
+        &registrar_cookie,
+        &voter_weight_record_cookie,
+        &voter_cookie,
+        &[&nft_cookie]
+    ).await?;
+
+    let err = nft_voter_test
+        .update_voter_weight_record(
+            &registrar_cookie,
+            &mut voter_weight_record_cookie,
+            VoterWeightAction::CreateGovernance,
+            &[&nft_weight_record_cookies[0], &nft_weight_record_cookies[0]]
+        ).await
+        .err()
+        .unwrap();
+
+    assert_nft_voter_err(err, NftVoterError::DuplicatedNftDetected);
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_update_voter_weight_with_same_cnft_error() -> Result<(), TransportError> {
     let mut nft_voter_test = NftVoterTest::start_new().await;
     let realm_cookie = nft_voter_test.governance.with_realm().await?;
     let registrar_cookie = nft_voter_test.with_registrar(&realm_cookie).await?;
